+++ mitmproxy/tools/web/webaddons.py
@@ -35,7 +35,9 @@ class WebAuth:
             "Password to protect the mitmweb user interface. "
             "Values starting with `$` are interpreted as an argon2 hash, "
             "everything else is considered a plaintext password. "
-            "If no password is provided, a random token is generated on startup.",
+            "If no password is provided, a random token is generated on startup."
+            "For automated calls, you can pass the password as token query parameter"
+            "or as X-Token header.",
         )
 
     def configure(self, updated) -> None: